+++ landingzones/views.py
@@ -835,7 +835,7 @@ class ZoneMoveView(
             irods_backend = get_backend_api('omics_irods')
             path = irods_backend.get_path(zone)
             with irods_backend.get_session() as irods:
-                stats = irods_backend.get_object_stats(irods, path)
+                stats = irods_backend.get_stats(irods, path)
                 if stats['file_count'] == 0:
                     messages.info(request, ZONE_MOVE_NO_FILES)
                     return redirect(